backend/requirements.txt
flask
pandas
Flask-SQLAlchemy
mysql-connector-python
flask-cors
PyJWT
bcrypt
